[level 2] Title: 더 맵게, Time: 652.66 ms, Memory: 140 MB -BaekjoonHub

+++ 프로그래머스/2/42626. 더 맵게/더 맵게.java
@@ -4,21 +4,21 @@ class Solution {
     public int solution(int[] scoville, int K) {
         int answer = 0;
         
-        PriorityQueue<Long> q = new PriorityQueue<>();
+        PriorityQueue<Integer> q = new PriorityQueue<>();
         
         for(int i : scoville){
-            q.offer((long) i);
+            q.offer(i);
         }
         
         while(!q.isEmpty() && q.peek() < K){
             answer++;
             
-            if (q.size() < 2) {
+            if(q.size() < 2){
                 return -1;
             }
             
-            long first = q.poll();
-            long second = q.poll();
+            int first = q.poll();
+            int second = q.poll();
             
             q.offer(first + (second * 2));
         }
